cleanup Path-classes
