+++ game/src/contrib/item/concreteItem/ItemResourceEgg.java
@@ -8,7 +8,7 @@ import core.Entity;
 import core.Game;
 import core.components.PositionComponent;
 import core.utils.components.draw.Animation;
-import core.utils.components.draw.SimpleIPath;
+import core.utils.components.path.SimpleIPath;
 
 import java.io.IOException;
