+++ game/src/core/utils/components/draw/CoreAnimationPriorities.java
@@ -1,8 +1,12 @@
 package core.utils.components.draw;
 
 /**
- * CoreAnimationPriorities contain animations which usually all entities have. which contain simple
- * idle animations, movement animations, and the default animations.
+ * Priorities for the core animations.
+ *
+ * <p>Priorities define the order in which animations will be drawn by the {@link
+ * core.systems.DrawSystem}
+ *
+ * @see CoreAnimations
  */
 public enum CoreAnimationPriorities {
     IDLE(1000),
@@ -11,10 +15,20 @@ public enum CoreAnimationPriorities {
 
     private final int priority;
 
+    /**
+     * Create an enum-value with the specified priority.
+     *
+     * @param priority The priority value for the animation.
+     */
     CoreAnimationPriorities(int priority) {
         this.priority = priority;
     }
 
+    /**
+     * Gets the priority value.
+     *
+     * @return The priority value.
+     */
     public int priority() {
         return priority;
     }
