+++ game/test/contrib/components/InventoryComponentTest.java
@@ -9,7 +9,7 @@ import contrib.item.Item;
 import core.Entity;
 import core.Game;
 import core.utils.components.draw.Animation;
-import core.utils.components.draw.SimpleIPath;
+import core.utils.components.path.SimpleIPath;
 
 import org.junit.After;
 import org.junit.Test;
