+++ game/test/core/utils/components/draw/AnimationTest.java
@@ -4,6 +4,9 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+import core.utils.components.path.IPath;
+import core.utils.components.path.SimpleIPath;
+
 import org.junit.Test;
 
 import java.util.List;
